code-samples/Python/trustyworthy-agents/requirements.txt
azure-ai-inference
dotenv